requirements.txt
six==1.11.0
numpy==1.15.4
scikit_image==0.14.0
pydicom==1.2.1
matplotlib==2.2.2
scipy==1.1.0
tensorflow==1.12.0
psutil==5.4.6
absl_py==0.6.1
Pillow==5.4.1
absl==0.0
skimage==0.0
